+++ src/power_framework/core/metrics/__init__.py
@@ -1,0 +1,1 @@
+"""Make ``power_framework.core.metrics`` a package."""